main.go and modes.go: 	Reorganize routines to make code testable, before PR you can't reuse routines to test them because there is no way to import from main package, now you can
