+++ libs/workloads.go
@@ -1,4 +1,4 @@
-package main
+package libs
 
 import (
 	"log"
