+++ modes.go
@@ -1,19 +1,14 @@
 package main
 
 import (
-	"bytes"
-	"crypto/sha256"
-	"encoding/binary"
 	"fmt"
 	"log"
-	"math/rand"
 	"strings"
 	"sync/atomic"
 	"time"
 
-	"github.com/codahale/hdrhistogram"
 	"github.com/gocql/gocql"
-	"github.com/pkg/errors"
+	. "github.com/scylladb/scylla-bench/libs"
 )
 
 type RateLimiter interface {
@@ -56,85 +51,12 @@ func NewRateLimiter(maximumRate int, timeOffset time.Duration) RateLimiter {
 	return &MaximumRateLimiter{period, time.Now(), 0}
 }
 
-type Result struct {
-	Final          bool
-	ElapsedTime    time.Duration
-	Operations     int
-	ClusteringRows int
-	Errors         int
-	Latency        *hdrhistogram.Histogram
-}
-
-type MergedResult struct {
-	Time                    time.Duration
-	Operations              int
-	ClusteringRows          int
-	OperationsPerSecond     float64
-	ClusteringRowsPerSecond float64
-	Errors                  int
-	Latency                 *hdrhistogram.Histogram
-}
-
-func NewMergedResult() *MergedResult {
-	result := &MergedResult{}
-	result.Latency = NewHistogram()
-	return result
-}
-
-func (mr *MergedResult) AddResult(result Result) {
-	mr.Time += result.ElapsedTime
-	mr.Operations += result.Operations
-	mr.ClusteringRows += result.ClusteringRows
-	mr.OperationsPerSecond += float64(result.Operations) / result.ElapsedTime.Seconds()
-	mr.ClusteringRowsPerSecond += float64(result.ClusteringRows) / result.ElapsedTime.Seconds()
-	mr.Errors += result.Errors
-	if measureLatency {
-		dropped := mr.Latency.Merge(result.Latency)
-		if dropped > 0 {
-			log.Print("dropped: ", dropped)
-		}
-	}
-}
-
-func NewHistogram() *hdrhistogram.Histogram {
-	if !measureLatency {
-		return nil
-	}
-	return hdrhistogram.New(time.Microsecond.Nanoseconds()*50, (timeout + timeout*2).Nanoseconds(), 3)
-}
-
 func HandleError(err error) {
 	if atomic.SwapUint32(&stopAll, 1) == 0 {
 		log.Print(err)
 		fmt.Println("\nstopping")
 		atomic.StoreUint32(&stopAll, 1)
 	}
-}
-
-func MergeResults(results []chan Result) (bool, *MergedResult) {
-	result := NewMergedResult()
-	final := false
-	for i, ch := range results {
-		res := <-ch
-		if !final && res.Final {
-			final = true
-			result = NewMergedResult()
-			for _, ch2 := range results[0:i] {
-				res = <-ch2
-				for !res.Final {
-					res = <-ch2
-				}
-				result.AddResult(res)
-			}
-		} else if final && !res.Final {
-			for !res.Final {
-				res = <-ch
-			}
-		}
-		result.AddResult(res)
-	}
-	result.Time /= time.Duration(concurrency)
-	return final, result
 }
 
 func RunConcurrently(maximumRate int, workload func(id int, resultChannel chan Result, rateLimiter RateLimiter)) *MergedResult {
@@ -160,11 +82,11 @@ func RunConcurrently(maximumRate int, workload func(id int, resultChannel chan R
 		}(i)
 	}
 
-	final, result := MergeResults(results)
+	final, result := MergeResults(results, timeout, measureLatency, concurrency)
 	for !final {
 		result.Time = time.Now().Sub(startTime)
-		PrintPartialResult(result)
-		final, result = MergeResults(results)
+		PrintPartialResult(result, errorRecordingLatency, measureLatency)
+		final, result = MergeResults(results, timeout, measureLatency, concurrency)
 	}
 	return result
 }
@@ -179,8 +101,8 @@ func NewResultBuilder() *ResultBuilder {
 	rb.FullResult = &Result{}
 	rb.PartialResult = &Result{}
 	rb.FullResult.Final = true
-	rb.FullResult.Latency = NewHistogram()
-	rb.PartialResult.Latency = NewHistogram()
+	rb.FullResult.Latency = NewHistogram(timeout, measureLatency)
+	rb.PartialResult.Latency = NewHistogram(timeout, measureLatency)
 	return rb
 }
 
@@ -206,7 +128,7 @@ func (rb *ResultBuilder) IncErrors() {
 
 func (rb *ResultBuilder) ResetPartialResult() {
 	rb.PartialResult = &Result{}
-	rb.PartialResult.Latency = NewHistogram()
+	rb.PartialResult.Latency = NewHistogram(timeout, measureLatency)
 }
 
 func (rb *ResultBuilder) RecordLatency(start time.Time, end time.Time) {
@@ -285,133 +207,13 @@ func RunTest(resultChannel chan Result, workload WorkloadGenerator, rateLimiter 
 	resultChannel <- *rb.FullResult
 }
 
-const (
-	generatedDataHeaderSize int64 = 24
-	generatedDataMinSize int64 = generatedDataHeaderSize + 33
-)
-
-func GenerateData(pk int64, ck int64, size int64) []byte {
-	if !validateData {
-		return make([]byte, size);
-	}
-
-	buf := new(bytes.Buffer)
-
-	if size < generatedDataHeaderSize {
-		binary.Write(buf, binary.LittleEndian, int8(size))
-		binary.Write(buf, binary.LittleEndian, pk ^ ck)
-	} else {
-		binary.Write(buf, binary.LittleEndian, size)
-		binary.Write(buf, binary.LittleEndian, pk)
-		binary.Write(buf, binary.LittleEndian, ck)
-		if size < generatedDataMinSize {
-			for i := generatedDataHeaderSize; i < size; i++ {
-				binary.Write(buf, binary.LittleEndian, int8(0))
-			}
-		} else {
-			payload := make([]byte, size - generatedDataHeaderSize - sha256.Size)
-			rand.Read(payload)
-			csum := sha256.Sum256(payload)
-			binary.Write(buf, binary.LittleEndian, payload)
-			binary.Write(buf, binary.LittleEndian, csum)
-		}
-	}
-
-	value := make([]byte, size)
-	copy(value, buf.Bytes())
-	return value
-}
-
-func ValidateData(pk int64, ck int64, data []byte) error {
-	if !validateData {
-		return nil
-	}
-
-	buf := bytes.NewBuffer(data)
-	size := int64(buf.Len())
-
-	var storedSize int64
-	if size < generatedDataHeaderSize {
-		var storedSizeCompact int8
-		err := binary.Read(buf, binary.LittleEndian, &storedSizeCompact)
-		if err != nil {
-			return errors.Wrap(err, "failed to validate data, cannot read size from value")
-		}
-		storedSize = int64(storedSizeCompact)
-	} else {
-		err := binary.Read(buf, binary.LittleEndian, &storedSize)
-		if err != nil {
-			return errors.Wrap(err, "failed to validate data, cannot read size from value")
-		}
-	}
-
-	if size != storedSize {
-		return errors.Errorf("actual size of value (%d) doesn't match size stored in value (%d)", size, storedSize)
-	}
-
-	// There is no random payload for sizes < minFullSize
-	if size < generatedDataMinSize {
-		expectedBuf := GenerateData(pk, ck, size)
-		if !bytes.Equal(buf.Bytes(), expectedBuf) {
-			return errors.Errorf("actual value doesn't match expected value:\nexpected: %x\nactual: %x", expectedBuf, buf.Bytes())
-		}
-		return nil
-	}
-
-	var storedPk, storedCk int64
-	var err error
-
-	// Validate pk
-	err = binary.Read(buf, binary.LittleEndian, &storedPk)
-	if err != nil {
-		return errors.Wrap(err, "failed to validate data, cannot read pk from value")
-	}
-	if storedPk != pk {
-		return errors.Errorf("actual pk (%d) doesn't match pk stored in value (%d)", pk, storedPk)
-	}
-
-	// Validate ck
-	err = binary.Read(buf, binary.LittleEndian, &storedCk)
-	if err != nil {
-		return errors.Wrap(err, "failed to validate data, cannot read pk from value")
-	}
-	if storedCk != ck {
-		return errors.Errorf("actual ck (%d) doesn't match ck stored in value (%d)", ck, storedCk)
-	}
-
-	// Validate checksum over the payload
-	payload := make([]byte, size - generatedDataHeaderSize - sha256.Size)
-	err = binary.Read(buf, binary.LittleEndian, payload)
-	if err != nil {
-		return errors.Wrap(err, "failed to verify checksum, cannot read payload from value")
-	}
-
-	calculatedChecksumArray := sha256.Sum256(payload)
-	calculatedChecksum := calculatedChecksumArray[0:]
-
-	storedChecksum := make([]byte, 32)
-	err = binary.Read(buf, binary.LittleEndian, storedChecksum)
-	if err != nil {
-		return errors.Wrap(err, "failed to verify checksum, cannot read checksum from value")
-	}
-
-	if !bytes.Equal(calculatedChecksum, storedChecksum) {
-		return errors.New(fmt.Sprintf(
-			"corrupt checksum or data: calculated checksum (%x) doesn't match stored checksum (%x) over data\n%x",
-			calculatedChecksum,
-			storedChecksum,
-			payload))
-	}
-	return nil
-}
-
 func DoWrites(session *gocql.Session, resultChannel chan Result, workload WorkloadGenerator, rateLimiter RateLimiter) {
 	query := session.Query("INSERT INTO " + keyspaceName + "." + tableName + " (pk, ck, v) VALUES (?, ?, ?)")
 
 	RunTest(resultChannel, workload, rateLimiter, func(rb *ResultBuilder) (error, time.Duration) {
 		pk := workload.NextPartitionKey()
 		ck := workload.NextClusteringKey()
-		value := GenerateData(pk, ck, clusteringRowSizeDist.Generate())
+		value := GenerateData(pk, ck, clusteringRowSizeDist.Generate(), validateData)
 		bound := query.Bind(pk, ck, value)
 
 		requestStart := time.Now()
@@ -440,7 +242,7 @@ func DoBatchedWrites(session *gocql.Session, resultChannel chan Result, workload
 		for !workload.IsPartitionDone() && atomic.LoadUint32(&stopAll) == 0 && batchSize < rowsPerRequest {
 			ck := workload.NextClusteringKey()
 			batchSize++
-			value := GenerateData(currentPk, ck, clusteringRowSizeDist.Generate())
+			value := GenerateData(currentPk, ck, clusteringRowSizeDist.Generate(), validateData)
 			batch.Query(request, currentPk, ck, value)
 		}
 
@@ -542,7 +344,7 @@ func DoReadsFromTable(table string, session *gocql.Session, resultChannel chan R
 			for iter.Scan(&resPk, &resCk, &value) {
 				rb.IncRows()
 				if validateData {
-					err := ValidateData(resPk, resCk, value)
+					err := ValidateData(resPk, resCk, value, validateData)
 					if err != nil {
 						rb.IncErrors()
 						log.Printf("data corruption in pk(%d), ck(%d): %s", resPk, resCk, err)
